Add files via upload
The hotel industry is experiencing a digital transformation, and mobile apps have become a focal point for
improving customer experience and engagement. In response to this trend, our hotel chain embarked on a
mission to develop and launch a mobile app, with the aim of increasing customer spending and enhancing overall
profitability.