Add files via upload
CMS contracted with CORE to work in collaboration with other contractors to develop the methodology for the Overall Hospital Quality Star Ratings on Hospital Compare. Hospital Compare includes information on over 100 quality measures and more than 4,000 hospitals. The primary objective of the Overall Hospital Quality Star Ratings project is to summarize information from the existing measures on Hospital Compare in a way that is useful and easy to interpret for patients and consumers through the development of a statistically sound Star Rating methodology. Consistent with other CMS Star Rating programs, this methodology assigns each hospital between one and five stars, reflecting the hospital’s overall performance on selected quality measures.